Amended setup description
Found out that upload of results is not scripted in NIcalc yet. Bård Pedersen got started on it, but it's not completed yet. We will look more into this when he comes back from holidays. As an alternative option, it should be possible to manually update tables in the database (Siw).